test: splitted and added tests
