+++ src/index.ts
@@ -1,7 +1,2 @@
-export * from './fileHeader';
-export * from './newDataBlock';
-export * from './oldDataBlock';
-export * from './logBlock';
-export * from './types';
-export * from './utility';
+export { guessSpectraType } from './utility/guessSpectraType';
 export * from './parse';
