+++ src/utility/__tests__/guessSpectraType.test.ts
@@ -1,0 +1,49 @@
+import { readFileSync } from 'fs';
+import { join } from 'path';
+
+import {IOBuffer} from "iobuffer";
+
+import { fileHeader } from '../../fileHeader';
+import { guessSpectraType } from '../guessSpectraType';
+
+
+describe('Test the Spectra-Type Guess', () => {
+
+const dataDir = "../../__tests__/data";
+  it('nir', () => {
+    //kubelka-monk is now in "other" as it is a 
+    //very different type of "ir"
+    const buffer = readFileSync(join(__dirname, dataDir, 'nir.spc'));
+    const result = fileHeader(new IOBuffer(buffer));
+    const guessedType = guessSpectraType(result);
+    expect(guessedType).toBe('other');
+  });
+
+
+  it('ft-ir.spc', () => {
+    const buffer = readFileSync(join(__dirname, dataDir, 'Ft-ir.spc'));
+    const result = fileHeader(new IOBuffer(buffer));
+    const guessedType = guessSpectraType(result);
+    expect(guessedType).toBe('ir'); //MIR I think (middle range IR.)
+  });
+
+  it('raman-sion.spc', () => {
+    const buffer = readFileSync(join(__dirname, dataDir, 'raman-sion.spc'));
+    const result = fileHeader(new IOBuffer(buffer));
+    const guessedType = guessSpectraType(result);
+    expect(guessedType).toBe('raman');
+  });
+
+  it('NDR0002.SPC', () => {
+    const buffer = readFileSync(join(__dirname, dataDir, 'NDR0002.SPC'));
+    const result = fileHeader(new IOBuffer(buffer));
+    const guessedType = guessSpectraType(result);
+    expect(guessedType).toBe('raman');
+  });
+  it('m_xyxy', () => {
+    const buffer = readFileSync(join(__dirname, dataDir, 'm_xyxy.spc'));
+    const result = fileHeader(new IOBuffer(buffer));
+    const guessedType = guessSpectraType(result);
+    expect(guessedType).toBe('mass');
+  });
+});
